# static/swagger.json
{
  "definitions": {
    "Exercise": {
      "type": "object",
      "properties": {
        "description": {
          "type": "string",
          "example": "Upper body strenght exercise"
        },
        "difficulty": {
          "type": "integer",
          "example": 2
        },
        "exercise_id": {
          "type": "integer",
          "example":  5
        },
        "instructions": {
          "type": "string",
          "example": "1. Lie on a flat bench with a barbell above your chest 2. Lower the barbell on your chest 3. Press the barbel back up to the starting position"
        },
        "name": {
          "type": "string",
          "example": "Bench Press"
        },
        "target_muscles": {
          "type": "string",
          "example": "Shoulders, Chest, Triceps"
        }
      }
    }
  },
  "info": {
    "description": "This API empowers you to design, track, and manage your personalized workout routines.",
    "termsOfService": "https://github.com/ggagua/workout-api/blob/main/README.md",
    "title": "Workout API",
    "version": "1.0.0"
  },
  "paths": {
    "/exercises": {
      "get": {
        "description": "<br/>",
        "responses": {
          "200": {
            "description": "List of exercises.",
            "schema": {
              "properties": {
                "exercises": {
                  "items": {
                    "$ref": "#/definitions/Exercise"
                  },
                  "type": "array"
                }
              },
              "type": "object"
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Retrieve all exercises."
      }
    },
    "/exercises/{exercise_id}": {
      "get": {
        "description": "<br/>",
        "parameters": [
          {
            "description": "ID of the exercise.",
            "in": "path",
            "name": "exercise_id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "The requested exercise.",
            "schema": {
              "$ref": "#/definitions/Exercise"
            }
          },
          "404": {
            "description": "Exercise not found.",
            "schema": {
              "properties": {
                "message": {
                  "example": "Exercise not found",
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Retrieve a specific exercise by ID."
      }
    },
    "/fitness-goals": {
      "get": {
        "description": "<br/>",
        "parameters": [
          {
            "description": "Bearer token for user authentication.",
            "in": "header",
            "name": "Authorization",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "List of fitness goals.",
            "schema": {
              "items": {
                "properties": {
                  "current_weight": {
                    "example": 70.5,
                    "type": "number"
                  },
                  "exercise_goals": {
                    "items": {
                      "properties": {
                        "name": {
                          "example": "Bench Press",
                          "type": "string"
                        },
                        "reps": {
                          "example": 10,
                          "type": "integer"
                        },
                        "reps_goal": {
                          "example": 12,
                          "type": "integer"
                        },
                        "weight": {
                          "example": 50.0,
                          "type": "number"
                        },
                        "weight_goal": {
                          "example": 60.0,
                          "type": "number"
                        }
                      },
                      "type": "object"
                    },
                    "type": "array"
                  },
                  "goal_weight": {
                    "example": 65.0,
                    "type": "number"
                  }
                },
                "type": "object"
              },
              "type": "array"
            }
          },
          "404": {
            "description": "Fitness goals not found.",
            "schema": {
              "properties": {
                "message": {
                  "example": "Fitness goals not found",
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Get all fitness goals for the authenticated user."
      },
           "post": {
        "description": "<br/>",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "type": "string",
            "required": true,
            "description": "Bearer token for user authentication."
          },
          {
            "in": "body",
            "name": "Fitness Goal JSON",
            "required": true,
            "schema": {
              "properties": {
                "current_weight": {
                  "example": 70.5,
                  "type": "number"
                },
                "goal_weight": {
                  "example": 65.0,
                  "type": "number"
                },
                "exercises_goals": {
                  "items": {
                    "properties": {
                      "name": {
                        "example": "Bench Press",
                        "type": "string"
                      },
                      "reps": {
                        "example": 10,
                        "type": "integer"
                      },
                      "reps_goal": {
                        "example": 12,
                        "type": "integer"
                      },
                      "weight": {
                        "example": 50.0,
                        "type": "number"
                      },
                      "weight_goal": {
                        "example": 60.0,
                        "type": "number"
                      }
                    },
                    "type": "object"
                  },
                  "type": "array"
                }
              },
              "type": "object"
            },
            "description": "JSON containing fitness goal information."
          }
        ],
        "responses": {
          "200": {
            "description": "Fitness goal created successfully.",
            "schema": {
              "properties": {
                "message": {
                  "example": "Fitness goal created successfully!",
                  "type": "string"
                }
              },
              "type": "object"
            }
          },
          "400": {
            "description": "Missing required goal information or exercise goal information.",
            "schema": {
              "properties": {
                "message": {
                  "example": "Missing required goal information or exercise goal information",
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Create a new fitness goal."
      }
  },
    "/fitness-goals/{goal_id}": {
      "delete": {
        "description": "<br/>",
        "parameters": [
          {
            "description": "Bearer token for user authentication.",
            "in": "header",
            "name": "Authorization",
            "required": true,
            "type": "string"
          },
          {
            "description": "ID of the fitness goal to be deleted.",
            "in": "path",
            "name": "goal_id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Fitness goal deleted successfully.",
            "schema": {
              "properties": {
                "message": {
                  "example": "Fitness goal deleted successfully",
                  "type": "string"
                }
              },
              "type": "object"
            }
          },
          "404": {
            "description": "Fitness goal not found.",
            "schema": {
              "properties": {
                "message": {
                  "example": "Fitness goal not found",
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Delete a fitness goal for the authenticated user."
      },
      "get": {
        "description": "<br/>",
        "parameters": [
          {
            "description": "Bearer token for user authentication.",
            "in": "header",
            "name": "Authorization",
            "required": true,
            "type": "string"
          },
          {
            "description": "ID of the fitness goal.",
            "in": "path",
            "name": "goal_id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Details of the requested fitness goal.",
            "schema": {
              "properties": {
                "fitness_goal": {
                  "properties": {
                    "current_weight": {
                      "example": 70.5,
                      "type": "number"
                    },
                    "exercise_goals": {
                      "items": {
                        "properties": {
                          "name": {
                            "example": "Bench Press",
                            "type": "string"
                          },
                          "reps": {
                            "example": 10,
                            "type": "integer"
                          },
                          "reps_goal": {
                            "example": 12,
                            "type": "integer"
                          },
                          "weight": {
                            "example": 50.0,
                            "type": "number"
                          },
                          "weight_goal": {
                            "example": 60.0,
                            "type": "number"
                          }
                        },
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "goal_weight": {
                      "example": 65.0,
                      "type": "number"
                    }
                  },
                  "type": "object"
                }
              },
              "type": "object"
            }
          },
          "404": {
            "description": "Fitness goal not found.",
            "schema": {
              "properties": {
                "message": {
                  "example": "Fitness goal not found",
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Get a single fitness goal for the authenticated user."
      }
    },
    "/login": {
      "get": {
        "description": "<br/>",
        "responses": {
          "200": {
            "description": "Authentication successful.(Use Basic Auth)",
            "schema": {
              "properties": {
                "token": {
                  "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJ1c2VyX2lkIjoiMjM0NTY3ODktYmFkZS00YmQwLTllYWEtYjg5Zjg0Y2NmZDEzIiwiZXhwIjoxNjQ1MzA2ODgxfQ.UvoEwVtwACuN6_EOMbF68kZOmGq2GQr5uobVGwWED18",
                  "type": "string"
                }
              },
              "type": "object"
            }
          },
          "401": {
            "description": "Authentication failed.",
            "schema": {
              "example": "Could not verify",
              "type": "string"
            }
          }
        },
        "security": [
          {
            "basicAuth": []
          }
        ],
        "summary": "User login endpoint."
      }
    },
    "/logout": {
      "post": {
        "description": "<br/>",
        "responses": {
          "200": {
            "description": "Logout successful.",
            "schema": {
              "properties": {
                "message": {
                  "description": "A success message.",
                  "example": "Logout successful",
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Logout the current user immediately."
      }
    },
...
    "/workout-plans/{workout_plan_id}": {
      "delete": {
        "description": "<br/>",
        "parameters": [
          {
            "description": "Bearer token for user authentication.",
            "in": "header",
            "name": "Authorization",
            "required": true,
            "type": "string"
          },
          {
            "description": "ID of the workout plan to be deleted.",
            "in": "path",
            "name": "workout_plan_id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Workout plan deleted successfully.",
            "schema": {
              "properties": {
                "message": {
                  "example": "Workout plan deleted successfully",
                  "type": "string"
                }
              },
              "type": "object"
            }
          },
          "404": {
            "description": "Workout plan not found.",
            "schema": {
              "properties": {
                "message": {
                  "example": "Workout plan not found",
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Delete a workout plan for the authenticated user."
      }
    }
  },
  "swagger": "2.0"
}